copy colors from vanilla food icon colors
We can retrieve the colors used in vanilla for food icons (the small
fork in the top right corner of foods in the inventory) and use them as
the basis for our own colors. This way, the colors are now consistent
and also we might get lucky on updates and have the colors automagically
update if food icons are updated.

Didn't manage to find where the vanilla alpha value for the HUD was
defined but we can at least manually set it up to be the exact same as
in vanilla, as seen by inspecting the HUD with UnityExplorer.